app/requirements.txt
# Usage: pip install -r requirements.txt

pandas
python-decouple
flask
boto3